Move ObjectArray
