+++ src/test/java/net/minestom/server/utils/collection/ObjectArrayTest.java
@@ -1,4 +1,4 @@
-package net.minestom.server.utils;
+package net.minestom.server.utils.collection;
 
 import org.junit.jupiter.api.Test;
